Components: Refactor MenuItem components
