+++ components/index.js
@@ -20,9 +20,9 @@ export { default as FormToggle } from './form-toggle';
 export { default as FormTokenField } from './form-token-field';
 export { default as IconButton } from './icon-button';
 export { default as KeyboardShortcuts } from './keyboard-shortcuts';
-export { default as MenuItemsChoice } from './menu-items/choice';
-export { default as MenuItemsGroup } from './menu-items/group';
-export { default as MenuItemsItem } from './menu-items/item';
+export { default as MenuGroup } from './menu-group';
+export { default as MenuItem } from './menu-item';
+export { default as MenuItemsChoice } from './menu-items-choice';
 export { default as ScrollLock } from './scroll-lock';
 export { NavigableMenu, TabbableContainer } from './navigable-container';
 export { default as Notice } from './notice';
