+++ components/menu-group/style.scss
@@ -1,0 +1,9 @@
+.components-menu-group {
+	width: 100%;
+	padding: 10px;
+}
+
+.components-menu-group__label {
+	margin-bottom: 10px;
+	color: $dark-gray-300;
+}
