+++ edit-post/components/header/more-menu/style.scss
@@ -21,7 +21,7 @@
 }
 
 .edit-post-more-menu__content {
-	.components-menu-items__group:not(:last-child) {
+	.components-menu-group:not(:last-child) {
   		border-bottom: 1px solid $light-gray-500;
 	}
 }
